refactor: split `expression.ts` into smaller piecies, feat: add `color` function support to the style language
