+++ src/core/mantle/evaluator/simple/expression/functions/index.ts
@@ -1,0 +1,3 @@
+export { binaryFunctions, getEvaluateBinaryFunction } from "./binaryFunctions";
+export { unaryFunctions, getEvaluateUnaryFunction } from "./unaryFunctions";
+export { Color } from "./colors";
